AG-12437 Move expand/collapse buttons to ApiReference
